Move classes to pipeline-scenario module
